pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-deepseek>=1.0.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
